+++ examples/pdl1-rfd/run-local-full.sh
@@ -42,6 +42,7 @@ nextflow run ${PIPELINE_DIR}/main.nf  \
   --refold_target_fasta='input/full/3BIK_B.fasta' \
   --refold_target_templates='input/full/' \
   --output_rmsd_aligned=true \
+  --enable_gpu_stats=true \
   -profile local \
   -resume \
   -with-report results/logs/report_${DATESTAMP}.html \